Refactor: Relocate TriggerXPermissionFlagManager and improve dialog
This commit relocates `TriggerXPermissionFlagManager` to a new file `TriggerXManualPermissionStatusManager.kt` and renames it accordingly.

Additionally, it enhances the permission guidance dialog in `TriggerXPermissionComposable.kt`:
- The dialog is no longer dismissible by clicking outside.
- A "Later" button has been added to the dismiss action.

Key changes:

- **triggerx/src/main/java/com/meticha/triggerx/permission/TriggerXPermissionComposable.kt**:
    - Replaced `TriggerXPermissionFlagManager` with `TriggerXManualPermissionStatusManager`.
    - Set `dismissOnClickOutside` to `false` in `DialogProperties` for the permission guidance dialog.
    - Added a `dismissButton` with "Later" text to the permission guidance dialog.
- **triggerx/src/main/java/com/meticha/triggerx/permission/TriggerXPermissionManager.kt**:
    - Replaced `TriggerXPermissionFlagManager` with `TriggerXManualPermissionStatusManager`.
    - Removed unused `OverlayPermissionGuidanceDialog` composable.
- **triggerx/src/main/java/com/meticha/triggerx/preference/TriggerXAlarmIdManager.kt**:
    - Removed the `TriggerXPermissionFlagManager` object.
- **triggerx/src/main/java/com/meticha/triggerx/preference/TriggerXManualPermissionStatusManager.kt**:
    - New file created.
    - Contains the relocated and renamed `TriggerXPermissionFlagManager` (now `TriggerXManualPermissionStatusManager`).
